requirements.txt
python-dotenv==0.16.0
eth-brownie==1.19.2
web3==5.31.1